rename files DashboardLayout, AuthLayout
