+++ frontend/src/layouts/DashboardLayout.vue
